New shiny features added to makefile. Effectively no bugs to report, and does not break previous uses.
If the makefile is changed, the idea that a rebuild is needed is far from fantastical, and therefore it is listed as a dependency for all object files.

if !IGNORE is found anywhere (by "find" + grep) in the path or filename, it shall be excluded, sparing
"clean-all" to clean dependency and object files there.

The "headers" file directly in src is included for every c file.
You may still have whatever header setup you had before, but this way you won't have to
confuse the directory structure if many unrelated modules use the same header; Just put it as global,
under a directory related to what the header does rather than what the modules that use it do.