using sys/time.h instead of time.h for accuracy reasons.
This does make it less portable, but it's pretty much a requirement for the precise timing required
for SDL applicatons, which this should support.

There are a couple options in the long-term for both portability and meeting this requirement:
1. use an external library for timing
2. recreate the functionality of such external library that I'll actually need, at the risk of reinventing the wheel
This doesn't seem like it should be too hard, but I'm not particularly interested in portability at the moment since it's just me using it.
But it's worth being wary about this and similar changes in the future.