Added some dynamic state + removed compiler errors + small changes
It's been a very long time since I last worked on this project, so I probably can't enumerate the changes very well, but it goes something like this:
- The viewport and scissor are now dynamic. This way, there's no need to recreate the whole pipeline whenever the swapchain needs to be recreated.
- Useless unused variable Timer, which I've just now noticed, added. I'm far too lazy to go back and remove that right now. Enjoy.
- Calloc's put the right way round, not that it seems to matter too much.
- The extension finding process was made a bit easier. See, the old way tried to load any extensions it could, and perhaps logged the ones it couldn't. The way I see it, if you're loading extensions you'd better need them.
- Semaphores actually work now. Before they were giving a validation layer error, because I was using one set when I should have been using two.
- Huge swaths of comments containing old, unused code were added to make navigating the codebase feel a little more like sifting through a wreckage
- Don't really remember how cube texture or shader was changed. But it appears they were so. There's some detective work to do.

Somewhere along the way it stopped crashing when it was resized too much. Something to do with memory leakages during the recreation of everything before, I think.
Shame that essential MVP feature has been dropped for the current iteration. I'll try to get it back to crashing ASAP, thank you.